requirements.txt
bs4
html2text
sentence-transformers
psycopg2-binary